face_detection and Photo Editing app/requirements.txt
opencv-python==4.1.2.30
numpy>=1.18.1
streamlit==0.61.0
pillow==7.1.2